@@ -29,33 +29,32 @@
   },
   "homepage": "https://github.com/alielkhateeb/mui-image-slider#readme",
   "peerDependencies": {
-    "@material-ui/core": "^4.0",
-    "react": "^16.8.0",
-    "react-dom": "^16.8.0"
+    "@mui/material": "^5.x",
+    "react": "^18.x",
+    "react-dom": "^18.x"
   },
   "devDependencies": {
     "@babel/cli": "^7.7.7",
     "@babel/core": "^7.7.7",
     "@babel/plugin-external-helpers": "^7.7.4",
-    "@babel/preset-env": "^7.7.7",
-    "@babel/preset-react": "^7.7.4",
-    "@material-ui/core": "^4.8.0",
+    "@babel/preset-env": "^7.23.2",
+    "@babel/preset-react": "^7.22.15",
+    "@mui/material": "^5.14.15",
     "@rollup/plugin-commonjs": "^11.0.0",
     "@rollup/plugin-replace": "^2.3.0",
-    "babel-jest": "^24.9.0",
+    "@testing-library/react": "^14.0.0",
+    "babel-jest": "^29.7.0",
     "babel-loader": "^8.0.6",
     "chai": "^4.2.0",
     "coveralls": "^3.0.9",
     "cross-env": "^6.0.3",
     "css-loader": "^6.8.1",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.2",
     "html-webpack-plugin": "^5.5.3",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "react-test-renderer": "^16.12.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-test-renderer": "^18.2.0",
     "rollup": "^1.27.14",
     "rollup-plugin-babel": "^4.3.3",
     "rollup-plugin-uglify": "^6.0.4",
